created new module clientAnimal.js to handle circular dependency issues. Linked animals and clients together, so deleting a client deletes their animals (and more), and creaed restrictions, like can't create an animal without linking a client in the database.

+++ client.js
@@ -1,17 +1,9 @@
-const mongoose = require("mongoose");
+const { Client, deleteClientAnimals } = require('./clientAnimal');
 
-const clientSchema = new mongoose.Schema({
-  firstName: String,
-  lastName: String,
-  email: String,
-  phoneNumber: String,
-  username: String,
-  password: String,
-});
-
-const client = mongoose.model("Clients", clientSchema);
+const client = Client;
 
 function setup(app) {
+
   app.post("/saveClient", (req, res) => {
     const newClient = new client({
       firstName: req.body.firstName,
@@ -20,6 +12,7 @@ function setup(app) {
       phoneNumber: req.body.phoneNumber,
       username: req.body.username,
       password: req.body.password,
+      animals: []
     });
 
     newClient
@@ -64,15 +57,26 @@ function setup(app) {
     try {
       // Extract the username from the request body
       const user = req.body.username;
-
-      // Use the findOneAndDelete method to remove the client based on the username
-      const result = await client.findOneAndDelete({ username: user });
-
-      if (result) {
-        console.log(
-          "Client Deleted: " + result.firstName + " " + result.lastName
-        );
-        res.status(200).send("Client deleted");
+  
+      // Find the client to get the list of associated animal IDs
+      const clientToDelete = await client.findOne({ username: user });
+  
+      if (clientToDelete) {
+        const animalIdsToDelete = clientToDelete.animals;
+  
+        // Use the deleteClientAnimals function to delete the associated animals
+        await deleteClientAnimals(animalIdsToDelete);
+  
+        // Now, delete the client
+        const result = await client.findOneAndDelete({ username: user });
+  
+        if (result) {
+          console.log("Client Deleted: " + result.firstName + " " + result.lastName + "\nTheir Animals Deleted");
+          res.status(200).send("Client and associated animals deleted");
+        } else {
+          console.log("Client not found");
+          res.status(404).send("Client not found");
+        }
       } else {
         console.log("Client not found");
         res.status(404).send("Client not found");
@@ -82,6 +86,9 @@ function setup(app) {
       res.status(500).send("Error Deleting client");
     }
   });
+  
+
+
 
   app.post('/updateClient', (req, res) => {
     try {
@@ -106,5 +113,5 @@ function setup(app) {
 }
 
 module.exports = {
-  setup: setup,
+  setup: setup
 };
